move ViewModel initialization to screen parameters
